+++ CombineParameters/(Page2)UserWindowParameters.xaml
@@ -1,0 +1,97 @@
+﻿<Window x:Class="CombineParameters.UserWindowParameters"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:CombineParameters"
+             mc:Ignorable="d" Height="445" Width="741"
+             Title="Combine parameters" Background="#FFCBCBCB">
+
+
+    <Grid RenderTransformOrigin="0.494,0.534">
+
+
+
+        <!--<ScrollViewer Margin="0,0,279,45" HorizontalAlignment="Right" Width="169" Height="160" VerticalAlignment="Bottom" >
+
+            <StackPanel 
+                    
+                    x:Name="stackPanel"
+                    Height="auto"
+                    HorizontalAlignment="Right" 
+                    VerticalAlignment="Top"
+                    Width="auto" RenderTransformOrigin="0.5,0.5"/>
+
+        </ScrollViewer>-->
+
+
+
+
+        <TextBlock
+                   HorizontalAlignment="Left" 
+                   TextWrapping="Wrap" 
+                   Text="Please select one or more categories:"
+                   Margin="10,10,0,688.2"
+                   Width="350" FontSize="20" />
+        <DataGrid x:Name="selectedParameters" HorizontalAlignment="Left" Height="221" Margin="10,99,0,0" VerticalAlignment="Top" Width="289"/>
+        <DataGrid x:Name="availableParametersToSelect2" HorizontalAlignment="Left" Height="221" Margin="374,99,0,0" VerticalAlignment="Top" Width="289"/>
+        <Label Content="Select those parameters you want to concatenate" HorizontalAlignment="Left" Margin="10,15,0,0" VerticalAlignment="Top" RenderTransformOrigin="-1.021,0.406"/>
+        <Label Content="Available parameters" HorizontalAlignment="Left" Margin="10,68,0,0" VerticalAlignment="Top" RenderTransformOrigin="-1.021,0.406"/>
+        <Label Content="Selected Parameters" HorizontalAlignment="Left" Margin="374,68,0,0" VerticalAlignment="Top" RenderTransformOrigin="-1.021,0.406"/>
+        <Button Content="+" HorizontalAlignment="Left" Margin="319,144,0,0" VerticalAlignment="Top" Width="35" Click="AddParameter"/>
+        <Button Content="-" HorizontalAlignment="Left" Margin="319,168,0,0" VerticalAlignment="Top" Width="35" Click="DeleteParameter"/>
+        <Button Content="Finish" HorizontalAlignment="Left" Margin="588,355,0,22" Width="75"/>
+        <Button Content="Go back" HorizontalAlignment="Left" Margin="10,355,0,22" Width="75"/>
+        <Button Content="Up" HorizontalAlignment="Left" Margin="374,331,0,0" VerticalAlignment="Top" Width="35" Click="UpButton"/>
+        <Button Content="Down" HorizontalAlignment="Left" Margin="414,331,0,0" VerticalAlignment="Top" Width="35" Click="DownButton"/>
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+        <!--<Button
+            x:Name="but"
+            Margin="204,105,224,137">
+            <Image Source="arrow.png" Height="35" Width="38"/>
+
+        </Button>
+
+        -->
+        <!--<Button
+            
+            Margin="204,170,224,75">
+            <Image Source="arrow.png" Height="31" Width="37" RenderTransformOrigin="0.5,0.5">
+                <Image.RenderTransform>
+                    <TransformGroup>
+                        <ScaleTransform/>
+                        <SkewTransform/>
+                        <RotateTransform Angle="-179.478"/>
+                        <TranslateTransform/>
+                    </TransformGroup>
+                </Image.RenderTransform>
+            </Image>
+
+        </Button>-->
+    </Grid>
+
+
+
+
+
+
+</Window>
