--- CombineParameters/UserWindow.xaml
@@ -1,67 +1,0 @@
-﻿<Window x:Class="CombineParameters.UserWindow"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:CombineParameters"
-             mc:Ignorable="d" Height="559.5" Width="590"
-             Title="Combine parameters" Background="#FFCBCBCB">
-
-    <Grid
-          OpacityMask="#FFC73E3E" Margin="0,0,2,6">
-
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="34"/>
-            <ColumnDefinition Width="552"/>
-            <ColumnDefinition Width="407"/>
-
-        </Grid.ColumnDefinitions>
-
-        <Grid.RowDefinitions>
-
-            <RowDefinition Height="84"/>
-            <RowDefinition Height="347"/>
-            <RowDefinition Height="1400"/>
-        </Grid.RowDefinitions>
-
-
-
-        <ScrollViewer Margin="0,0,61,32" Grid.Row="1" Grid.Column="1" HorizontalAlignment="Right" Width="473" Height="284" VerticalAlignment="Bottom" >
-
-            <StackPanel 
-                    
-                    x:Name="stackPanel"
-                    Height="auto"
-                    HorizontalAlignment="Right" 
-                    VerticalAlignment="Top"
-                    Width="auto" RenderTransformOrigin="0.5,0.5"/>
-
-        </ScrollViewer>
-
-        <Button 
-            
-                Content="Next" 
-                VerticalAlignment="Top" 
-                Margin="401,33,61,0"
-                Click="Button_Click" Height="30" Grid.Column="1" Grid.Row="2"/>
-
-        <Button 
-                Grid.Column="1"
-                Grid.Row="2"
-                Content="Cancel"
-                HorizontalAlignment="Left" 
-                VerticalAlignment="Top" 
-                Width="90"
-                Margin="57,33,0,0"
-                Click="ButtonCancelClick" Height="30" />
-
-        <TextBlock 
-                   Grid.Row="0" 
-                   TextWrapping="Wrap" 
-                   Text="Please select one or more categories:"
-                   Margin="18,22,28,26" FontSize="28" Grid.Column="1" />
-
-
-
-    </Grid>
-</Window>
