--- CombineParameters/UserWindowParameters.xaml.cs
@@ -1,34 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
-using Autodesk.Revit.DB;
-
-
-
-namespace CombineParameters
-{
-    /// <summary>
-    /// Interaction logic for UserWindowParameters.xaml
-    /// </summary>
-    public partial class UserWindowParameters : Window
-    {
-        //List<Parameter> _parameters; 
-        public UserWindowParameters()
-        //public UserWindowParameters(List<Parameter> parameters)
-        {
-            InitializeComponent();
-            //_parameters = parameters;
-        }
-    }
-}
